% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{changeShoworder-ViSibook-method}
\alias{changeShoworder}
\alias{changeShoworder,ViSibook,numeric-method}
\alias{changeShoworder-ViSibook-method}
\title{Method \code{changeShoworder}}
\usage{
changeShoworder(book, v)

\S4method{changeShoworder}{ViSibook,numeric}(book, v)
}
\arguments{
\item{book}{a ViSibook object.}

\item{v}{a vector containing the number from the slot "showorder" of book.}
}
\value{
book a ViSibook object.
}
\description{
The method \code{changeShoworder} allows the slot "showorder" of a ViSibook to be modified.
}
\seealso{
\code{\linkS4class{ViSibook}} and see \code{\link{plot-ViSigrid-method}} for examples.
}

