% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shworder_from_data.R
\name{convert_visibook_graph}
\alias{convert_visibook_graph}
\title{Function \code{convert_visibook_graph}}
\usage{
convert_visibook_graph(X, book, Xsup = NULL, typeg = "weigthed_vertice")
}
\arguments{
\item{X}{a data.frame containing execution times}

\item{book}{a ViSiBook}

\item{Xsup}{a data.frame containing supplementary times}

\item{typeg}{weighted_vertice : Only one vertices exists between two action weigthed by the number of individuals having pass through
vertices : One vertive is created for each passage}
}
\value{
a ViSibook object.
}
\description{
convert_visibook_graph Convert a visibook into a graph
}
\seealso{
See \code{\linkS4class{ViSibook}} to get the definitions of the columns
and see  \code{\link{plot-ViSigrid-method}} for examples.
}

