% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{ConvertFromViSibook-ViSibook-method}
\alias{ConvertFromViSibook-ViSibook-method}
\alias{ConvertFromViSibook,ViSibook-method}
\alias{ConvertFromViSibook}
\title{Method \code{ConvertFromViSibook-ViSibook}}
\usage{
ConvertFromViSibook(x)

\S4method{ConvertFromViSibook}{ViSibook}(x)
}
\arguments{
\item{x}{a ViSibook object.}
}
\value{
a data.frame.
}
\description{
The method \code{ConvertFromViSibook} converts a ViSibook in a data.frame object.
}
\seealso{
\code{\linkS4class{ViSibook}} and see \code{\link{plot-ViSigrid-method}} for examples.
}
