% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{dim-ViSibook-method}
\alias{dim-ViSibook-method}
\alias{dim,ViSibook-method}
\title{Method \code{dim-ViSibook}}
\usage{
\S4method{dim}{ViSibook}(x)
}
\arguments{
\item{x}{a ViSibook object.}
}
\value{
Vector \itemize{
\item{[1]}{ The number of actions defined in x.}
\item{[2]}{ The number of characteristics defined in x,
its minimum value is 6 and its maximum is 15.}
}
}
\description{
Method Dim for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}
}
