% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{[,ViSibook,numeric,missing,ANY-method}
\alias{[,ViSibook,numeric,missing,ANY-method}
\alias{[,ViSibook,numeric,missing-method}
\alias{[,ViSibook,missing,numeric,ANY-method}
\alias{[,ViSibook,missing,numeric-method}
\alias{[,ViSibook,numeric,numeric,ANY-method}
\alias{[,ViSibook,numeric,numeric-method}
\title{Method get for ViSibook object.}
\usage{
\S4method{[}{ViSibook,numeric,missing,ANY}(x, i, j, drop = TRUE)

\S4method{[}{ViSibook,missing,numeric,ANY}(x, i, j, drop = TRUE)

\S4method{[}{ViSibook,numeric,numeric,ANY}(x, i, j, drop = TRUE)
}
\arguments{
\item{x}{a ViSibook object.}

\item{i}{a numeric.}

\item{j}{a numeric.}

\item{drop}{= TRUE.}
}
\value{
obj.
}
\description{
Method get for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}.
}
