% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ViSibook.R
\docType{methods}
\name{plot-ViSibook-method}
\alias{plot-ViSibook-method}
\alias{plot,ViSibook-method}
\title{Method \code{plot-ViSibook}}
\usage{
\S4method{plot}{ViSibook}(x, ncharmax = 10, ncharmaxdelay = 50)
}
\arguments{
\item{x}{a ViSibook object.}

\item{ncharmax}{is the maximum number of plotted character for the labels of punctual actions, set to 10.}

\item{ncharmaxdelay}{is the maximum number of plotted character for the labels of long actions, set to 50.}
}
\description{
Method plot for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}, \code{\link{visielse}}
}
