% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.Visibook.r
\docType{methods}
\name{print,ViSibook-method}
\alias{print,ViSibook-method}
\alias{print,ViSibook-methods}
\alias{print}
\alias{print-ViSibook-methods}
\title{Method \code{print-ViSibook}}
\usage{
\S4method{print}{ViSibook}(x)
}
\arguments{
\item{x}{a ViSibook object.}
}
\description{
Method print for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}},  \code{\link{visielse}}, 
and see \code{\link{plot-ViSigrid-method}} for examples.
}
