% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shworder_from_data.R
\name{root_visibook}
\alias{root_visibook}
\title{Function \code{root_visibook}}
\usage{
root_visibook(g, book)
}
\arguments{
\item{g}{a graph}

\item{book}{a data.frame corresponding to a ViSibook}
}
\value{
a ViSibook object.
}
\description{
root_visibook select a root for a graph
}

