% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{set-ViSibook-method}
\alias{set-ViSibook-method}
\alias{[<-,ViSibook,numeric,numeric,ANY-method}
\alias{[<-,ViSibook,missing,numeric,ANY-method}
\alias{[<-,ViSibook,numeric,missing,ANY-method}
\title{Method set for ViSibook object.}
\usage{
\S4method{[}{ViSibook,numeric,numeric,ANY}(x, i, j) <- value

\S4method{[}{ViSibook,missing,numeric,ANY}(x, i, j) <- value

\S4method{[}{ViSibook,numeric,missing,ANY}(x, i, j) <- value
}
\arguments{
\item{x}{a ViSibook object.}

\item{i}{a numeric.}

\item{j}{a numeric.}

\item{value}{object to allocate.}
}
\value{
a ViSibook object.
}
\description{
Method set for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}
}
