% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{methods}
\name{show-ViSibook-method}
\alias{show-ViSibook-method}
\alias{show,ViSibook-method}
\title{Method \code{show-ViSibook}}
\usage{
\S4method{show}{ViSibook}(object)
}
\arguments{
\item{object}{a ViSibook .}
}
\description{
Method show for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}.
}
