% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shworder_from_data.R
\name{showorder_from_data}
\alias{showorder_from_data}
\title{Function \code{showorder_from_data}}
\usage{
showorder_from_data(X, book = NULL, Xsup = NULL, idsubject = 1)
}
\arguments{
\item{X}{a data.frame containing execution times}

\item{book}{a ViSiBook}

\item{Xsup}{a data.frame containing supplemetary times when there is}

\item{idsubject}{integer 1}
}
\value{
a ViSibook object.
}
\description{
showorder_from_data Extract the showorder from the data
}
\seealso{
See \code{\linkS4class{ViSibook}} to get the definitions of the columns
and see  \code{\link{plot-ViSigrid-method}} for examples.
}

