% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ViSigrid.r
\docType{methods}
\name{summary-ViSigrid-method}
\alias{summary-ViSigrid-method}
\alias{summary,ViSigrid-method}
\title{Method \code{summary-ViSigrid}}
\usage{
\S4method{summary}{ViSigrid}(object)
}
\arguments{
\item{object}{a ViSigrid.}
}
\value{
list \itemize{
\item{ \strong{ punctual} }{ summary of punctual actions (typeA=="p").}
\item{ \strong{ longs} }{  summary of long actions (typeA=="p"). }
}
}
\description{
Method summary for ViSigrid object.
}
\seealso{
\code{\linkS4class{ViSigrid}}, \code{\link{visielse}},\code{\linkS4class{ViSibook}}.
and see \code{\link{plot-ViSigrid-method}} for examples.
}
