% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibook.r
\docType{class}
\name{ViSibook-class}
\alias{ViSibook-class}
\alias{ViSibook}
\title{Class \code{ViSiBook}}
\description{
Class \code{ViSibook} defines the structure of the process to be plotted.
}
\section{Slots}{

\describe{
\item{\code{vars}}{a vector storing names of actions.}

\item{\code{label}}{a vector storing brief description of actions.}

\item{\code{typeA}}{a vector storing type of actions, "l" for long actions ( which have a stating time and an ending time ), "p" for punctual actions.}

\item{\code{showorder}}{a vector storing order in which actions will be plotted. When an action should not be plotted,
\code{showorder} should be \code{NA}.}

\item{\code{deb}}{a vector. \itemize{
\item{Long actions}{ \code{deb} stores the punctual action names that delimit the long actions beginning.}
\item{Punctual action}{ \code{NA} .}
}}

\item{\code{fin}}{a vector. \itemize{
\item{Long actions}{ \code{fin} stores the punctual action names that delimit the long actions ending.}
\item{Punctual actions}{ \code{NA} .}
}}

\item{\code{GZDeb}}{a vector, optional, \code{GZdeb} stores punctual actions green zone starting time.}

\item{\code{GZFin}}{a vector, optional, \code{GZFin} stores punctual actions green zone ending time.}

\item{\code{Repetition}}{optional a vector, optional, When a green zone is defined,
\code{Repetition} stores the duration of the time interval between green zones.}

\item{\code{BZBeforeDeb}}{a vector, optional, \code{BZBeforeDeb} a vector storing punctual black zone 1 starting time.}

\item{\code{BZBeforeFin}}{a vector, optional, \code{BZBeforeFin} storing punctual black zone 1 ending time.}

\item{\code{BZAfterDeb}}{a vector, optional, \code{BZAfterDeb} stores punctual black zone 2 starting time .}

\item{\code{BZAfterFin}}{a vector, optional, \code{BZAfterFin} stores punctual black zone 2 ending time.}

\item{\code{BZLong}}{a vector, optional, \code{BZLong} stores the long action black zone time.}

\item{\code{BZLtype}}{a vector, optional, \code{BZLtype} stores the type of the black zone,
"time" if the action should be finish at a time, "span" if the action should be finish in a time.}

\item{\code{NAMES}}{a vector storing names of slots that are to be defined.}
}}

\seealso{
\code{\link{visielse}} for examples.
}
