% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibookfromDATA.R
\name{ViSibookfromDATA}
\alias{ViSibookfromDATA}
\title{Function \code{ViSibookfromDATA}}
\usage{
ViSibookfromDATA(X, idsubject = 1)
}
\arguments{
\item{X}{data.frame.}

\item{idsubject}{numeric indicates the number of the column of X which stores id.}
}
\value{
a ViSibook corresponding to the dataset X.
}
\description{
\code{ViSibookfromDATA} build an object class ViSibook from observational data. The process is the ordered list of punctual actions given by the columns names of X.
}
