% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSigrid.r
\docType{class}
\name{ViSigrid-class}
\alias{ViSigrid-class}
\alias{ViSigrid}
\title{Class \code{ViSigrid}}
\description{
Class \code{ViSigrid} defines the structure of the process to be plotted.
}
\section{Slots}{

\describe{
\item{\code{MATp}}{A \code{"dgCMatrix"}. It stores the grid for all punctual actions in the book.}

\item{\code{MATpsup}}{A \code{"dgCMatrix"}. It stores the grid for all punctual actions 
in the book corresponding to the supplementary times.}

\item{\code{idsup}}{A \code{"vector"} It stores individuals id having supplementary times.}

\item{\code{colvect}}{A \code{"matrix"} Matrix with colors to use.}

\item{\code{L}}{A \code{"data.frame"} It stores the data corresponding to long actions having a showorder.}

\item{\code{idsort}}{A \code{"matrix"} For all long actions, it stores the order of individuals in which each actions will be plot.}

\item{\code{BZL}}{A \code{"dgCMatrix"} It stores black zones for long actions, calculated for each individuals.}

\item{\code{Lsup}}{A \code{"data.frame"} for the long actions having  a showorder and supplementary times defined, it 
stores the data corresponding to those actions.}

\item{\code{book}}{A \code{"ViSibook"} it stores the structure of the grid for the plot.}

\item{\code{group}}{A \code{"factor"} it stores the group for the each individuals.}

\item{\code{vect_tps}}{A \code{"vector"} it stores the times vector mapping the grid.}

\item{\code{informers}}{A \code{"matrix"} It stores the statistics (mean, median or NULL) by actions.}

\item{\code{testsP}}{A \code{"vector"} Results of tests p.value<threshold.test.}

\item{\code{parameters}}{A \code{"list"}. It stores the parameters put in the \code{\link{visielse}} function.}
}}

\seealso{
\code{\link{visielse}}, \code{\link{plot,ViSigrid-method}}, \code{\linkS4class{ViSibook}}
}
