% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.ViSibook.r
\docType{methods}
\name{initialize,ViSibook-method}
\alias{initialize,ViSibook-method}
\title{Method \code{initialize-ViSibook}}
\usage{
\S4method{initialize}{ViSibook}(.Object, vars, label, typeA, showorder, deb,
  fin, GZDeb, GZFin, Repetition, BZBeforeDeb, BZBeforeFin, BZAfterDeb,
  BZAfterFin, BZLong, BZLtype, NAMES)
}
\arguments{
\item{.Object}{a ViSibook object.}

\item{vars}{a vector storing names of actions.}

\item{label}{a vector. storing brief description of actions.}

\item{typeA}{Vector storing type of actions, "l" for long actions, "p" for punctuals.}

\item{showorder}{vector storing order in which actions will be plotted, is an actions is not to be plot its showorder is "NA".}

\item{deb}{Vector storing, for long actions, the punctual action names that corresponds to its start.}

\item{fin}{Vector storing, for long actions, the punctual action that corresponds to its end.}

\item{GZDeb}{Vector storing punctuals actions green zone starting time.}

\item{GZFin}{Vector storing punctual action green zone ending time.}

\item{Repetition}{Vector storing if the green zones should be repeated the time interval of repetition.}

\item{BZBeforeDeb}{Vector storing punctual black zone 1 starting time.}

\item{BZBeforeFin}{Vector storing punctual black zone 1 ending time.}

\item{BZAfterDeb}{Vector storing punctual black zone 2 starting time.}

\item{BZAfterFin}{Vector storing punctual black zone 2 ending time.}

\item{BZLong}{Vector storing the long action black zone time.}

\item{BZLtype}{Vector storing the type of the black zone,
"time" if the action should be finish at a time, "span" if the action should be finish in a time.}

\item{NAMES}{Vector storing names of slots that are to be considered for \code{\link{plot-ViSigrid-method}}.}
}
\value{
a ViSibook object
}
\description{
Method initialize for class ViSibook object.
}
\seealso{
See \code{\link{plot-ViSigrid-method}} for examples.
}

