% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ViSibook.R
\docType{methods}
\name{plot-ViSibook-method}
\alias{plot,ViSibook-method}
\alias{plot-ViSibook-method}
\title{Method \code{plot-ViSibook}}
\usage{
\S4method{plot}{ViSibook}(x, ncharmax = 10, ncharmaxdelay = 50)
}
\arguments{
\item{x}{a ViSibook object.}

\item{ncharmax}{is the number maximum of plotted character for the labels of punctual actions, set to 10.}

\item{ncharmaxdelay}{number maximum of  plotted character for the labels of long actions, set to 50.}
}
\description{
Method plot for ViSibook object.
}
\seealso{
\code{\linkS4class{ViSibook}}, \code{\link{buildViSiGrid}}, 
\code{\link{changeShoworder-ViSibook-method}} and see \code{\link{plot-ViSigrid-method}} for examples.
}

