% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ViSibookfromDATA.R
\name{ViSibookfromDATA}
\alias{ViSibookfromDATA}
\title{Function \code{ViSibookfromDATA}}
\usage{
ViSibookfromDATA(X, idsubject = 1)
}
\arguments{
\item{X}{data.frame}

\item{idsubject}{numeric number of the column which stores id}
}
\value{
a ViSiBook
}
\description{
The function \code{ViSibookfromDATA} build an object class ViSiBook from observational datas
}

